007: Move Core layer into separate Core project
