--- AlcoPlus.API/Configurations/Swagger/SwaggerOptions.cs
@@ -1,40 +1,0 @@
-﻿using Asp.Versioning.ApiExplorer;
-using Microsoft.Extensions.Options;
-using Microsoft.OpenApi.Models;
-using Swashbuckle.AspNetCore.SwaggerGen;
-
-namespace AlcoPlus.API.Configurations.Swagger;
-
-public class SwaggerOptions : IConfigureOptions<SwaggerGenOptions>
-{
-    private readonly IApiVersionDescriptionProvider _provider;
-
-    public SwaggerOptions(IApiVersionDescriptionProvider provider) => _provider = provider;
-
-    public void Configure(SwaggerGenOptions options)
-    {
-        foreach (var description in _provider.ApiVersionDescriptions)
-        {
-            options.SwaggerDoc(description.GroupName, CreateInfoForApiVersion(description));
-        }
-    }
-
-    private static OpenApiInfo CreateInfoForApiVersion(ApiVersionDescription description)
-    {
-        var info = new OpenApiInfo()
-        {
-            Title = "Example Web API",
-            Version = description.ApiVersion.ToString(),
-            Description = "Description for the example Web API",
-            Contact = new OpenApiContact { Name = "Author name", Email = "author-main@org.com" },
-            License = new OpenApiLicense { Name = "MIT", Url = new Uri("https://opensource.org/licenses/MIT") }
-        };
-
-        if (description.IsDeprecated)
-        {
-            info.Description += " This API version has been deprecated.";
-        }
-
-        return info;
-    }
-}
