--- AlcoPlus.API/Contracts/ICountriesRepository.cs
@@ -1,8 +1,0 @@
-﻿using AlcoPlus.Data;
-
-namespace AlcoPlus.API.Contracts;
-
-public interface ICountriesRepository : IRepository<Country>
-{
-    Task<Country> GetDetails(int id);
-}
