--- AlcoPlus.API/Contracts/IHotelsRepository.cs
@@ -1,7 +1,0 @@
-﻿using AlcoPlus.Data;
-
-namespace AlcoPlus.API.Contracts;
-
-public interface IHotelsRepository : IRepository<Hotel>
-{
-}
