+++ AlcoPlus.API/Controllers/AccountController.cs
@@ -1,5 +1,5 @@
-﻿using AlcoPlus.API.Contracts;
-using AlcoPlus.API.Models.Users;
+﻿using AlcoPlus.Core.Contracts;
+using AlcoPlus.Core.Models.Users;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
