+++ AlcoPlus.API/Controllers/v2/AccountController.cs
@@ -1,13 +1,12 @@
-﻿using AlcoPlus.API.Contracts;
-using AlcoPlus.API.Models.Users;
-using Asp.Versioning;
+﻿using AlcoPlus.Core.Contracts;
+using AlcoPlus.Core.Models.Users;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
 namespace AlcoPlus.API.Controllers.v2;
 
 [Route("api/v{version:apiVersion}/account")]
-[ApiVersion("2.0")]
+[Asp.Versioning.ApiVersion("2.0")]
 [ApiController]
 public class AccountController : ControllerBase
 {
