+++ AlcoPlus.API/Controllers/v2/CountriesController.cs
@@ -1,18 +1,17 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using AutoMapper;
-using AlcoPlus.API.Models.Countries;
-using AlcoPlus.API.Contracts;
+using AlcoPlus.Core.Models.Countries;
+using AlcoPlus.Core.Contracts;
 using Microsoft.AspNetCore.Authorization;
-using AlcoPlus.API.Exceptions;
-using Asp.Versioning;
+using AlcoPlus.Core.Exceptions;
 using Microsoft.AspNetCore.OData.Query;
 using AlcoPlus.Data;
 
 namespace AlcoPlus.API.Controllers.v2;
 
 [Route("api/v{version:apiVersion}/countries")]
-[ApiVersion("2.0")]
+[Asp.Versioning.ApiVersion("2.0")]
 [ApiController]
 [Authorize]
 public class CountriesController : ControllerBase
