--- AlcoPlus.API/Exceptions/NotFoundException.cs
@@ -1,8 +1,0 @@
-﻿namespace AlcoPlus.API.Exceptions;
-
-public class NotFoundException : ApplicationException
-{
-    public NotFoundException(string name, object key) : base($"{name} ({key}) was not found")
-    {
-    }
-}
