--- AlcoPlus.API/Models/Countries/CreateCountryDto.cs
@@ -1,7 +1,0 @@
-﻿using System.ComponentModel.DataAnnotations;
-
-namespace AlcoPlus.API.Models.Countries;
-
-public class CreateCountryDto : BaseCountryDto
-{
-}
