--- AlcoPlus.API/Models/Countries/GetCountryDto.cs
@@ -1,6 +1,0 @@
-﻿namespace AlcoPlus.API.Models.Countries;
-
-public class GetCountryDto : BaseCountryDto
-{
-    public int Id { get; set; }
-}
