--- AlcoPlus.API/Models/Countries/UpdateCountryDto.cs
@@ -1,6 +1,0 @@
-﻿namespace AlcoPlus.API.Models.Countries;
-
-public class UpdateCountryDto : BaseCountryDto
-{
-    public int Id { get; set; }
-}
