--- AlcoPlus.API/Models/Hotels/BaseHotelDto.cs
@@ -1,18 +1,0 @@
-﻿using System.ComponentModel.DataAnnotations;
-
-namespace AlcoPlus.API.Models.Hotels;
-
-public abstract class BaseHotelDto
-{
-    [Required]
-    public string Name { get; set; }
-
-    [Required]
-    public string Address { get; set; }
-
-    public double? Rating { get; set; }
-
-    [Required]
-    [Range(1, int.MaxValue)]
-    public int CountryId { get; set; }
-}
