--- AlcoPlus.API/Models/Hotels/CreateHotelDto.cs
@@ -1,5 +1,0 @@
-﻿namespace AlcoPlus.API.Models.Hotels;
-
-public class CreateHotelDto : BaseHotelDto
-{
-}
