--- AlcoPlus.API/Models/Hotels/HotelDto.cs
@@ -1,6 +1,0 @@
-﻿namespace AlcoPlus.API.Models.Hotels;
-
-public class HotelDto : BaseHotelDto
-{
-    public int Id { get; set; }
-}
