--- AlcoPlus.API/Models/Hotels/UpdateHotelDto.cs
@@ -1,6 +1,0 @@
-﻿namespace AlcoPlus.API.Models.Hotels;
-
-public class UpdateHotelDto : BaseHotelDto
-{
-    public int Id { get; set; }
-}
