--- AlcoPlus.API/Models/Users/AuthResponseDto.cs
@@ -1,9 +1,0 @@
-﻿namespace AlcoPlus.API.Models.Users;
-
-public class AuthResponseDto
-{
-    public string UserId { get; set; }
-
-    public string Token { get; set; }
-    public string RefreshToken { get; set; }
-}
