--- AlcoPlus.API/Models/Users/LoginUserDto.cs
@@ -1,13 +1,0 @@
-﻿using System.ComponentModel.DataAnnotations;
-
-namespace AlcoPlus.API.Models.Users;
-
-public class LoginUserDto
-{
-    [Required]
-    [EmailAddress]
-    public string Email { get; set; }
-
-    [Required]
-    public string Password { get; set; }
-}
