+++ AlcoPlus.Core/Configurations/Swagger/SwaggerConfig.cs
@@ -1,0 +1,54 @@
+﻿using Microsoft.AspNetCore.Mvc.ApiExplorer;
+using Microsoft.AspNetCore.Mvc.ModelBinding;
+using Microsoft.OpenApi.Models;
+using Swashbuckle.AspNetCore.SwaggerGen;
+using System.Text.Json;
+
+namespace AlcoPlus.Core.Configurations.Swagger;
+
+public class SwaggerConfig : IOperationFilter
+{
+    public void Apply(OpenApiOperation operation, OperationFilterContext context)
+    {
+        var apiDescription = context.ApiDescription;
+        operation.Deprecated |= apiDescription.IsDeprecated();
+
+        foreach (var responseType in context.ApiDescription.SupportedResponseTypes)
+        {
+            var responseKey = responseType.IsDefaultResponse ? "default" : responseType.StatusCode.ToString();
+            var response = operation.Responses[responseKey];
+
+            foreach (var contentType in response.Content.Keys)
+            {
+                if (responseType.ApiResponseFormats.All(x => x.MediaType != contentType))
+                {
+                    response.Content.Remove(contentType);
+                }
+            }
+        }
+
+        if (operation.Parameters == null)
+        {
+            return;
+        }
+
+        foreach (var parameter in operation.Parameters)
+        {
+            var description = apiDescription.ParameterDescriptions.First(p => p.Name == parameter.Name);
+
+            parameter.Description ??= description.ModelMetadata?.Description;
+
+            if (parameter.Schema.Default == null &&
+                 description.DefaultValue != null &&
+                 description.DefaultValue is not DBNull &&
+                 description.ModelMetadata is ModelMetadata modelMetadata)
+            {
+                // REF: https://github.com/Microsoft/aspnet-api-versioning/issues/429#issuecomment-605402330
+                var json = JsonSerializer.Serialize(description.DefaultValue, modelMetadata.ModelType);
+                parameter.Schema.Default = OpenApiAnyFactory.CreateFromJson(json);
+            }
+
+            parameter.Required |= description.IsRequired;
+        }
+    }
+}
