+++ AlcoPlus.Core/Contracts/ICountriesRepository.cs
@@ -1,0 +1,8 @@
+﻿using AlcoPlus.Data;
+
+namespace AlcoPlus.Core.Contracts;
+
+public interface ICountriesRepository : IRepository<Country>
+{
+    Task<Country> GetDetails(int id);
+}
