+++ AlcoPlus.Core/Contracts/IHotelsRepository.cs
@@ -1,0 +1,7 @@
+﻿using AlcoPlus.Data;
+
+namespace AlcoPlus.Core.Contracts;
+
+public interface IHotelsRepository : IRepository<Hotel>
+{
+}
