+++ AlcoPlus.Core/Exceptions/NotFoundException.cs
@@ -1,0 +1,8 @@
+﻿namespace AlcoPlus.Core.Exceptions;
+
+public class NotFoundException : ApplicationException
+{
+    public NotFoundException(string name, object key) : base($"{name} ({key}) was not found")
+    {
+    }
+}
