+++ AlcoPlus.Core/Models/Countries/BaseCountryDto.cs
@@ -1,0 +1,11 @@
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace AlcoPlus.Core.Models.Countries;
+
+public abstract class BaseCountryDto
+{
+    [Required]
+    public string Name { get; set; }
+    public string ShortName { get; set; }
+
+}
