+++ AlcoPlus.Core/Models/Countries/CountryDto.cs
@@ -1,0 +1,10 @@
+﻿using AlcoPlus.Core.Models.Hotels;
+
+namespace AlcoPlus.Core.Models.Countries;
+
+public class CountryDto : BaseCountryDto
+{
+    public int Id { get; set; }
+
+    public List<HotelDto> Hotels { get; set; }
+}
