+++ AlcoPlus.Core/Models/Countries/CreateCountryDto.cs
@@ -1,0 +1,5 @@
+﻿namespace AlcoPlus.Core.Models.Countries;
+
+public class CreateCountryDto : BaseCountryDto
+{
+}
