+++ AlcoPlus.Core/Models/Countries/GetCountryDto.cs
@@ -1,0 +1,6 @@
+﻿namespace AlcoPlus.Core.Models.Countries;
+
+public class GetCountryDto : BaseCountryDto
+{
+    public int Id { get; set; }
+}
