+++ AlcoPlus.Core/Models/Countries/UpdateCountryDto.cs
@@ -1,0 +1,6 @@
+﻿namespace AlcoPlus.Core.Models.Countries;
+
+public class UpdateCountryDto : BaseCountryDto
+{
+    public int Id { get; set; }
+}
