+++ AlcoPlus.Core/Models/Hotels/CreateHotelDto.cs
@@ -1,0 +1,5 @@
+﻿namespace AlcoPlus.Core.Models.Hotels;
+
+public class CreateHotelDto : BaseHotelDto
+{
+}
