+++ AlcoPlus.Core/Models/Hotels/HotelDto.cs
@@ -1,0 +1,6 @@
+﻿namespace AlcoPlus.Core.Models.Hotels;
+
+public class HotelDto : BaseHotelDto
+{
+    public int Id { get; set; }
+}
