+++ AlcoPlus.Core/Models/Hotels/UpdateHotelDto.cs
@@ -1,0 +1,6 @@
+﻿namespace AlcoPlus.Core.Models.Hotels;
+
+public class UpdateHotelDto : BaseHotelDto
+{
+    public int Id { get; set; }
+}
