+++ AlcoPlus.Core/Models/Users/AuthResponseDto.cs
@@ -1,0 +1,9 @@
+﻿namespace AlcoPlus.Core.Models.Users;
+
+public class AuthResponseDto
+{
+    public string UserId { get; set; }
+
+    public string Token { get; set; }
+    public string RefreshToken { get; set; }
+}
