+++ AlcoPlus.Core/Repository/HotelsRepository.cs
@@ -1,0 +1,11 @@
+﻿using AlcoPlus.Core.Contracts;
+using AlcoPlus.Data;
+
+namespace AlcoPlus.Core.Repository;
+
+public class HotelsRepository : Repository<Hotel>, IHotelsRepository
+{
+    public HotelsRepository(AlcoPlusDbContext alcoPlusDbContext) : base(alcoPlusDbContext)
+    {
+    }
+}
